Fix celotool commands, all testnets working
