+++ packages/helm-charts/testnet/templates/proxy.service.yaml
@@ -1,7 +1,5 @@
 {{ range $validatorIndex, $proxyCount := .Values.geth.proxiesPerValidator }}
 
-{{ $proxyIpAddresses := (splitList "/" (index $.Values.geth.proxyIPAddressesPerValidatorArray $validatorIndex)) }}
-
 {{ range $index, $e := until ($proxyCount | int) }}
 
 {{ template "celo.node-service" (dict "Values" $.Values "Release" $.Release "Chart" $.Chart "index" $index "service_type" "ClusterIP" "svc_name" (print "validators-" $validatorIndex "-proxy-internal") "node_name" (print "validators-" $validatorIndex "-proxy") "name_suffix" "" "component_label" "proxy" "load_balancer_ip" "" ) }}
@@ -22,6 +20,7 @@
 {{ end }}
 ---
 {{ if $.Values.geth.static_ips }}
+{{ $proxyIpAddresses := (splitList "/" (index $.Values.geth.proxyIPAddressesPerValidatorArray $validatorIndex)) }}
 {{ $loadBalancerIP := index $proxyIpAddresses $index }}
 {{ template "celo.node-service" (dict "Values" $.Values "Release" $.Release "Chart" $.Chart "index" $index "service_type" "LoadBalancer" "node_name" (print "validators-" $validatorIndex "-proxy") "component_label" "proxy" "load_balancer_ip" $loadBalancerIP ) }}
   ports:
