+++ packages/terraform-modules/testnet/modules/full-node/outputs.tf
@@ -1,13 +1,13 @@
 output ip_addresses {
-  value = [ for v in google_compute_address.full_node : v.address ]
+  value = [for v in google_compute_address.full_node : v.address]
 }
 
 output internal_ip_addresses {
-  value = [ for v in google_compute_instance.full_node : v.network_interface.0.network_ip ]
+  value = [for v in google_compute_instance.full_node : v.network_interface.0.network_ip]
 }
 
 output self_links {
-  value = [ for v in google_compute_instance.full_node : v.self_link ]
+  value = [for v in google_compute_instance.full_node : v.self_link]
 }
 
 output ip_addresses_map {
